Moving the debug ops into their own package to clean up the main package.
